requirements.txt
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
setuptools==78.1.1
sqlparse==0.5.3
wheel==0.45.1
django-cors-headers==4.8.0